+++ src/sass/components/_movie-info.scss
@@ -171,7 +171,7 @@
     left: 50%;
     transform: translate(-50%, -50%);
 
-    max-width: 882px;
+    max-width: 900px;
     max-height: 570px;
     padding: 50px 30px 40px 30px;
   }
